backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.2.2
supabase==2.15.3
cryptography==45.0.6
apscheduler==3.11.0
notion-client==3.0.0
requests==2.32.5
